adjusting module imports so that tests and heroku script will both work
