+++ test/conftest.py
@@ -1,12 +1,18 @@
+
 import pytest
 import torch
 from multiprocessing import Process
-from app.pg_rest_api import create_app
 from flask_migrate import Migrate
 from flask_sqlalchemy import SQLAlchemy
 import os
 import tempfile
 
+# We need to add our rest api as a path since it is a separate application
+# deployed on Heroku:
+import sys, os
+myPath = os.path.dirname(os.path.abspath(__file__))
+sys.path.insert(0, myPath + '/../app/pg_rest_api')
+from app.pg_rest_api.pg_app import create_app
 
 import syft
 from syft import TorchHook
