+++ test/flask_base_test.py
@@ -1,7 +1,6 @@
 import unittest
-from app.pg_rest_api import create_app, db
-from app.pg_rest_api.models import Worker, WorkerObject
-
+from app.pg_rest_api.pg_app import create_app
+from app.pg_rest_api.lib.models import db, Worker, WorkerObject
 
 class FlaskBaseTestCase(unittest.TestCase):
     def setUp(self):
